% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importCNMF}
\alias{importCNMF}
\title{Import cNMF results then export to an NMF object}
\usage{
importCNMF(
  prefix = NULL,
  metrics = "score",
  minRank = NULL,
  maxRank = NULL,
  by = 1,
  denThre = NULL,
  cnmfDir = "NULL",
  verbose = TRUE
)
}
\arguments{
\item{prefix}{\code{character} The run_name parameter from cNMF run}

\item{metrics}{\code{character} Either score or TPM, default is score}

\item{minRank}{\code{numeric} The minimum rank from cNMF run}

\item{maxRank}{\code{numeric} the maximum rank from cNMF run}

\item{by}{\code{numeric} increment of the sequence}

\item{denThre}{\code{character} The density threshold in string with a underscore}

\item{cnmfDir}{\code{character} Directory to the cNMF results}
}
\value{
An NMF object
}
\description{
Import cNMF results then export to an NMF object
}
\examples{
nmfObj <- importCNMF(prefix = "PREFIX", metrics = "score", minRank = 2, maxRank = 10, by = 1, denThre = "2_0", cnmfDir = "./cNMF")

}
