% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeNMFObjs}
\alias{mergeNMFObjs}
\title{Merge multiple NMF objects into one}
\usage{
mergeNMFObjs(nmfObjL)
}
\arguments{
\item{nmfObjL}{\code{list} A list of NMF objects}
}
\value{
An NMF object
}
\description{
Merge multiple NMF objects into one
}
\examples{
nmfObj <- mergeNMFObjs(nmfObjL = objL)

}
