% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{plotCorrDensity}
\alias{plotCorrDensity}
\title{Generate correlation coefficient density plot}
\usage{
\S4method{plotCorrDensity}{SOSet}(
  object,
  filename = "pairwiseCorrDensityPlot.pdf",
  breaks = NULL,
  width = 15,
  height = 15
)
}
\arguments{
\item{object}{A \code{SOSet} object}

\item{filename}{\code{character} Output file name}

\item{width}{\code{numeric} Width size for output PDF}

\item{height}{\code{numeric} Height size for output PDF}
}
\value{
Coefficient density plot
}
\description{
Generate correlation coefficient density plot
}
\examples{
plotCorrDensity(SOSet)

}
