% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SOSet.R
\name{show,SOSet-method}
\alias{show,SOSet-method}
\title{Show a SOSet}
\usage{
\S4method{show}{SOSet}(object)
}
\arguments{
\item{object}{\code{SOSet}}
}
\value{
Prints the SOSet object to the output stream, and returns invisible NULL.
}
\description{
Show a SOSet
}
