% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SOTK.R
\name{show,SOTK-method}
\alias{show,SOTK-method}
\title{Show a SOTK}
\usage{
\S4method{show}{SOTK}(object)
}
\arguments{
\item{object}{\code{SOTK}}
}
\value{
Prints the SOTK object to the output stream, and returns invisible NULL.
}
\description{
Show a SOTK
}
