% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-SOTK.R
\docType{class}
\name{SOTK-class}
\alias{SOTK-class}
\alias{SOTK}
\title{SOTK object and constructor}
\usage{
SOTK(
  SOSet,
  coefThre = 0.5,
  seed = 123456,
  niter = 1000,
  drop = FALSE,
  searchMet = "greedy",
  commWeight = 100,
  cohortWeight = 10
)
}
\arguments{
\item{SOSet}{A \code{SOSet} of the NMF results in SOSet class.}

\item{coefThre}{A \code{numeric} of the coefficient threshold value to include node in the correlation network. Default is 0.5.}

\item{seed}{A \code{numeric} of the seed to generate a correlation network. Default is 1118.}

\item{niter}{A \code{numeric} of the number of iteration when the igraph package generates a correlation network. Default is 1000.}

\item{drop}{\code{logical} Whether drop small/isolated island(s) from the correlation network. Default is FALSE.}

\item{searchMet}{\code{character} of the community search method, either greedy (fast greedy), leiden, betweeness, randomwalk (random walk), eigen, or louvain. Default is greedy.}

\item{commWeight}{\code{numeric} of the weight value on community info. If users want to rearrange nodes that identified as the same community closer, set a higher value. Default is 100.}

\item{cohortWeight}{\code{numeric} of the weight value on data/cohort (input). If users want to rearrange nodes closer for each data/cohort, set a higher value. Default is 5.}
}
\value{
SOTK object
}
\description{
An S4 class to contain Mx input and output
}
\details{
A
}
