% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\docType{methods}
\name{statComm}
\alias{statComm}
\title{Generate community network stats}
\usage{
\S4method{statComm}{SOTK}(object, filename = "stats.pdf", width = 10, height = 10)
}
\arguments{
\item{object}{A \code{SOTK} object}

\item{filename}{\code{character} Output file name}

\item{width}{\code{numeric} Width size for output PDF}

\item{height}{\code{numeric} Height size for output PDF}
}
\value{
Stacked bar plot and number of Metagenes per community in each dataset
}
\description{
Generate community network stats
}
\examples{
statComm(object = SOTK)

}
